refactor(ScoreBoardResponse): dto 패키지를 domain 패키지 위쪽으로 변경
